requirement.txt
pytesseract
pillow
fastapi
uvicorn
python-multipart